Submodule testcase_compiler into dwarf_explore
I'd like to avoid carrying binaries around but would also like to keep
dwarf_explore separate from the testcase compiler (as that may be used
for other projects in the future and really falls outside the scope of
dwarf_explore).

So now we submodule the testcase compiler and rely on the user (or test
runner) to compile it with clang. The dwarf_explore tests can now point
to the compiled testcase in an expected location.