Parser: return containing CU on GetEntry
Most callers can ignore the new return argument.

CAVEAT: adding this many returns to GetEntry is fishy. Additionally, we
need to persist the last CU we saw, which we can't do with our current
function-only architecture. Adding an object to track CU during GetEntry
is quite a bit of work which we'll get to later (probably).