Rename package client to memoryview
Client was a poor name for the function of memoryview. The real purpose
of memoryview is to give durins-door a method for reading and writing
bytes out of some representation of memory. The simplest implementation
of this is reading and writing to a file but more exotic backends are
possible: potentially reading data directly from memory via pointer or
reading bytes from a remote device.

While clients may need to implement certain aspects of memoryview and
provide appropriate callbacks (not yet clear how this will happen),
overall the role of a client is much different because typically,
clients will drive the interaction - not the other way around.